Made a few modifications and tried to compile inside MAME running Borland Turbo C, added a command number input, adjusted column number to better fit the screen I saw in the emulator
I was, until this point, attempting to read the disk using the disk number
2, that made sense in my head, floppy A is 0, floppy B is 1, and hard disk
C is 2, however, the C drive is NOT necessarily the first hard disk, it might
not even contain the entire disk, since it represents a sort of "mounted"
partition.

Now, the code SO FAR read the contents of both floppy drives perfectly,
but when I tried to run it with diskNumber 2, it failed to read...

I searched online and stumbled upon a wikipedia article dealing with INT13H,
which is the basis of the functioning of the biosread function I was using,
and there I found a table telling me that de hard disk was actually diskNumber
80h, so I ran the program, plopped 128 in that sucker, and voila, read the
first sector like a dream.

TODO:

Loop for reading entire drive
Safeguards for evil commands like formatting and writing
Sending stuff to the serial port
Writing a companion software on linux to receive the information from the
  a serial port and transforming it into a workable image file (.img, .chd, etc)