Compiled and tested on actual hardware (Scopus Nexus 2600, Brazilian XT Clone), on borland`s turbo C 2.0, some changes had to be made for it to compile, those changes were transfered to here.
-// comments changed to /* */ comments
-bioscom was expecting something to actually read the data sent, and apparently
hangs if there isn`t anything conected to the COM1 port, commented out the
bioscom call and the software ran perfectly

Couldn't replicate this, but one time while running the screen stopped filling
up with the data being read, it was like someone kept pressing enter, but the
software could'nt keep up.

I'll run some tests and try to find out whats happening, and will also bring
a serial to usb adapter and try to take an actual image of the disk this time.